@@ -10,32 +10,43 @@
     "url": "https://github.com/fabnavi/fabnavi5/issues"
   },
   "dependencies": {
-    "flux": "^2.1.1",
-    "jquery": "^3.0.0",
-    "jquery-ujs": "^1.2.1",
+    "axios": "0.15.3",
+    "debug": "2.6.0",
     "keymirror": "^0.1.1",
-    "machina": "^2.0.0-1",
-    "react": "^0.14.8",
-    "react-router": "^2.4.0"
+    "react": "^15.4.1",
+    "react-addons-create-fragment": "^15.4.2",
+    "react-bootstrap-table": "^3.0.0",
+    "react-dom": "15.4.1",
+    "react-paginate": "^4.1.1",
+    "react-redux": "5.0.1",
+    "react-router": "3.0.0",
+    "redux": "3.6.0",
+    "redux-actions": "^1.2.0",
+    "redux-observable": "0.12.2",
+    "rxjs": "5.0.3"
   },
   "devDependencies": {
-    "react-jade": "git://github.com/edgemethod/react-jade.git#19db22e",
-    "react-jade-loader": "git://github.com/fabnavi/react-jade-loader.git#adb23dc",
-    "webpack": "^1.13.0",
-    "webpack-dev-server": "^1.14.1",
     "babel-core": "^6.7.7",
+    "babel-eslint": "^7.1.1",
     "babel-loader": "^6.2.4",
     "babel-plugin-dev-expression": "^0.2.1",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-polyfill": "^6.20.0",
+    "babel-preset-es2015": "^6.18.0",
     "babel-preset-es2015-loose": "^7.0.0",
     "babel-preset-es2015-loose-native-modules": "^1.0.0",
     "babel-preset-react": "^6.5.0",
     "babel-preset-stage-0": "^6.5.0",
-    "bower-webpack-plugin": "^0.1.9",
-    "json-loader": "^0.5.4",
-    "node-sass": "^3.6.0",
+    "babel-preset-stage-3": "^6.17.0",
     "esdoc": "^0.4.6",
-    "eslint": "^2.5.1"
+    "eslint": "^2.5.1",
+    "eslint-plugin-react": "^6.8.0",
+    "eslint-watch": "^2.1.14",
+    "json-loader": "^0.5.4",
+    "leasot": "^4.3.0",
+    "node-sass": "^3.6.0",
+    "webpack": "^1.13.0",
+    "webpack-dashboard": "^0.2.1",
+    "webpack-dev-server": "^1.14.1"
   },
   "license": "MIT",
   "engines": {
@@ -43,8 +54,10 @@
   },
   "scripts": {
     "start": "webpack --progress --colors --watch",
-    "build": "webpack",
-    "lint": "eslint -c ./.eslintrc app/assets/javascripts/client/*",
+    "dash": "webpack-dashboard -- webpack -d --watch",
+    "build": "NODE_ENV=production webpack",
+    "lint": "eslint -c ./.eslintrc --fix `git ls-files ./app/assets/javascripts/client`",
+    "lint:watch": "esw -c ./.eslintrc --watch --cache --color `git ls-files ./app/assets/javascripts/client`",
     "doc": "esdoc -c esdoc.json"
   }
 }
